+++ src/fheroes2/editor/editor_interface.cpp
@@ -667,10 +667,10 @@ namespace Interface
         }
     }
 
-    void EditorInterface::setObjectOnTile( Maps::Tiles & tile, const Maps::ObjectGroup group, const int objectType )
+    void EditorInterface::setObjectOnTile( Maps::Tiles & tile, const Maps::ObjectGroup group, const int32_t objectType )
     {
         const auto objectInfo = Maps::getObjectsByGroup( group );
-        if ( objectType < 0 || objectType >= objectInfo.size() ) {
+        if ( objectType < 0 || objectType >= static_cast<int32_t>( objectInfo.size() ) ) {
             return;
         }
